Allow subscriptions in management functions
Authored-By: Scott Prutton <scott@systeminit.com>
