+++ lib/dal/src/attribute.rs
@@ -1,6 +1,7 @@
 //! This module contains the "attribute" domain concept, which primarily involves
 //! [`AttributeValue`](crate::AttributeValue) and [`AttributePrototype`](crate::AttributePrototype).
 
+pub mod attributes;
 pub mod path;
 pub mod prototype;
 pub mod value;
